requirements.txt
boto3
psycopg2-binary
streamlit
panda
sqlalchemy
faker
python-dotenv
scikit-learn